requirements.txt
feedparser==6.0.10
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
streamlit==1.22.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.0 